Make block an expression, not statement
Blocks being expressions allow us to group expressions (like parentheses) and also will open the space not to need explicit returns in the future.

The implementation is a bit hacky on the parser side *and* the interpreter side, but it works. When the parser encounters a { it can't know yet if it is a block or a map, so it tries to parse a statement. If it can, then it's a block. If not, it rewinds back and tries to parse a map.

Because we still have statements inside blocks, the interpreter now has a field to tack the last expression evaluated. The block will return that expression.